#35  Feat: 네이버 로그인 및 로그아웃

+++ src/main/java/com/example/barrier_free/domain/user/dto/SignupRequest.java
@@ -1,6 +1,5 @@
 package com.example.barrier_free.domain.user.dto;
 
-import com.example.barrier_free.domain.user.enums.UserType;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -15,5 +14,5 @@ public class SignupRequest {
     private String password;
     private String verifyPassword;
     private String userType;
-    private List<Long> userFacilityIds;
+    private List<Integer> userFacilityIds;
 }
